trimmed mr
